+++ Minigin/LevelLoader.h
@@ -19,7 +19,7 @@ namespace dae
 		int GetIndex(std::shared_ptr<GameObject> object);
 		int GetIndex(Node* node);
 		int GetIndex(float x, float y);
-		std::shared_ptr<GameObject> CheckGrid(int index) { if (index >= NodeArray.size()) return nullptr; if (index < 0) return nullptr; return NodeArray[index]; }
+		std::shared_ptr<GameObject> CheckGrid(int index) { if (index >= (int)NodeArray.size()) return nullptr; if (index < 0) return nullptr; return NodeArray[index]; }
 		std::shared_ptr<GameObject> CheckGrid(float x, float y);
 		std::vector<ActorComponent*> GetDigDugChars() { return mDigDugChars; }
